+++ packages/runtime-handler/src/dev-runtime/internal/functionRunner.ts
@@ -6,7 +6,7 @@ import {
   constructGlobalScope,
   isTwiml,
 } from '../route';
-import { ServerConfig } from '../types';
+import { ServerConfig, Headers } from '../types';
 import { Response } from './response';
 import { setRoutes } from './route-cache';
 
@@ -16,7 +16,7 @@ const sendDebugMessage = (debugMessage: string, ...debugArgs: any) => {
 
 export type Reply = {
   body?: string | number | boolean | object;
-  headers?: { [key: string]: number | string };
+  headers?: Headers;
   statusCode: number;
 };