+++ packages/runtime-handler/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.2.0-rc.3](https://github.com/twilio-labs/serverless-toolkit/compare/@twilio/runtime-handler@1.2.0-rc.2...@twilio/runtime-handler@1.2.0-rc.3) (2021-08-03)
+
+**Note:** Version bump only for package @twilio/runtime-handler
+
+
+
+
+
+# [1.2.0-rc.2](https://github.com/twilio-labs/serverless-toolkit/compare/@twilio/runtime-handler@1.2.0-rc.1...@twilio/runtime-handler@1.2.0-rc.2) (2021-07-28)
+
+
+### Bug Fixes
+
+* **runtime-handler:** using set-cookie now sets cookie header ([#332](https://github.com/twilio-labs/serverless-toolkit/issues/332)) ([6f65bc3](https://github.com/twilio-labs/serverless-toolkit/commit/6f65bc3bb692b8bd0b21d932f66ae394000e51a9))
+
+
 # [1.2.0-beta.0](https://github.com/twilio-labs/serverless-toolkit/compare/@twilio/runtime-handler@1.1.3...@twilio/runtime-handler@1.2.0-beta.0) (2021-09-25)
 
 
@@ -20,6 +36,14 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+
+# [1.2.0-rc.1](https://github.com/twilio-labs/serverless-toolkit/compare/@twilio/runtime-handler@1.1.1...@twilio/runtime-handler@1.2.0-rc.1) (2021-07-14)
+
+
+### Features
+
+* **handler:** update header and cookie support for Response ([#296](https://github.com/twilio-labs/serverless-toolkit/issues/296)) ([e9ef02e](https://github.com/twilio-labs/serverless-toolkit/commit/e9ef02ed9e10635623f462db6f53de3669ffaf0b))
+* **runtime-handler:** handle incoming headers and cookies ([#293](https://github.com/twilio-labs/serverless-toolkit/issues/293)) ([62ff180](https://github.com/twilio-labs/serverless-toolkit/commit/62ff1801db6a121122fcd944a855ad7f038cafe4))
 
 
 ## [1.1.2](https://github.com/twilio-labs/serverless-toolkit/compare/@twilio/runtime-handler@1.1.1...@twilio/runtime-handler@1.1.2) (2021-07-19)